+++ src/com/tutorial/main/HUD.java
@@ -5,7 +5,7 @@ import java.awt.*;
 public class HUD {
 
     // Static because we only have one health (we do 100 because it is a good point value / most games use 100 for health)
-    public static int HEALTH = 100;
+    public static float HEALTH = 100;
 
     // Score you see on screen
     private int score = 0;
@@ -15,7 +15,7 @@ public class HUD {
 
     public void tick() {
         // And we clamp the health to not go outside of the bar at 0 while it decreases due to collisions
-        HEALTH = Game.clamp(HEALTH, 0, 100);
+        HEALTH = (int) Game.clamp(HEALTH, 0, 100);
         score++;
 
        /* greenValue = Game.clamp(greenValue, 0, 255);
@@ -30,7 +30,7 @@ public class HUD {
         // Sets color to get more red as health decreases
         g.setColor(Color.getHSBColor((1f * HEALTH) / 360, 1f, 1f));
         // the amount of health you have (We HEALTH * 2 for now because the background is 200 and health is 100)
-        g.fillRect(15, 15, HEALTH * 2, 32);
+        g.fillRect(15, 15, (int) (HEALTH * 2), 32);
         // draws a border around the box
         g.setColor(Color.white);
         g.drawRect(15, 15, 200, 32);
